ci: make fmt and clippy real blocking gates (not allow_failure)
Actually fix the ~59 pre-existing clippy errors and repo-wide fmt drift
instead of leaving them allow_failure, per explicit request. Workspace is
now fully `cargo fmt --all -- --check` and `cargo clippy --workspace
--all-targets -- -D warnings` clean; both gates are hard-blocking in CI
again.

Real fixes (not suppressions):
- 19 FFI entry points in elevate-pam/pam/src/ffi.rs, plus ~35 pam_sm_*
  functions across the individual pam-* module crates: added accurate
  `# Safety` docs describing each function's actual pointer-validity
  contract (pamh provenance, argv/argc, out-pointers) instead of blanket
  #[allow]s.
- elevate/src/pam/dynload.rs: safety doc + comment for the dlsym call.
- 21 collapsible-if sites in elevate/ (exec/, elevators/,
  system/term/find_tty.rs) rewritten with let-chains (elevate/ is on
  edition 2024) -- same behavior, less nesting.
- 12 `.lines().flatten()` sites across elevate-umbra and elevate-pam
  switched to `.lines().map_while(Result::ok)`, clippy's own fix:
  `flatten()` silently loops forever on a repeating I/O error,
  `map_while` stops.
- Misc: redundant casts (delay.rs, limits.rs), a `let _ =` around a
  unit-returning macro (log.rs), a manual char-comparison closure
  (login_defs.rs), an unnecessary `.last()` on a DoubleEndedIterator
  (newgrp.rs), one leftover needless `return` (unix.rs).

Also, separately reported test failures now fixed for real:
- zoneinfo_path() only ever checked Zainium's own
  /overlayer/syshub/etc/zoneinfo -- never /usr/share/zoneinfo -- so
  installing tzdata in CI did nothing for test_tzinfo. Added
  /usr/share/zoneinfo as a portability fallback in
  elevate_paths::zoneinfo_dirs(), matching this crate's existing
  Zainium-first-then-FHS-fallback convention elsewhere.
- rust:latest runs CI jobs as literal root by default; several tests are
  real non-root-only privilege/permission checks and were failing purely
  because of that (confirmed by reproducing under `unshare --user
  --map-root-user`). Already fixed in the previous commit (run as an
  unprivileged user); this commit fixes the tzdata root cause fully.

Also included: in-progress pam_unix chauthtok work (see next commit) --
cargo fmt reformatted essentially the whole tree in one pass, so it
wasn't practical to separate that file's changes out cleanly.

+++ elevate-pam/pam/src/delay.rs
@@ -6,6 +6,6 @@ pub fn sleep_usec(usec: u32) {
         return;
     }
     let secs = (usec / 1_000_000) as u64;
-    let nanos = ((usec % 1_000_000) * 1000) as u32;
+    let nanos = (usec % 1_000_000) * 1000;
     std::thread::sleep(std::time::Duration::new(secs, nanos));
 }
